requirements.txt
# =============================
# Web / API
# =============================
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1

# =============================
# LangChain (STABLE SET)
# =============================
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.5
# =============================
# LLM Providers
# =============================
langchain-ollama==0.2.3
langchain-openai==0.3.3
# ❌ DO NOT PIN langchain-core

# =============================
# OpenAI SDK (compatibility only)
# =============================
openai>=1.58.1,<2.0.0
tiktoken==0.8.0

# =============================
# Vector Store / Embeddings
# =============================
faiss-cpu==1.9.0.post1
sentence-transformers==3.1.1
numpy==1.26.4
scipy==1.14.1

# =============================
# Database
# =============================
psycopg[binary]==3.2.3

# =============================
# Redis
# =============================
redis==5.0.8

# =============================
# Scraper / Document Parsing
# =============================
pymupdf==1.24.14
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.9.2
pydantic-settings>=2.4.0,<3.0.0
typing-extensions==4.12.2
detoxify==0.5.2

